+++ drf_tweaks/test_utils.py
@@ -10,21 +10,40 @@ class TooManySQLQueriesException(Exception):
     pass
 
 
-test_query_counter = 0
+class TestQueryCounter(object):
+    __instance = None
+
+    def __new__(cls, *args, **kwargs):
+        if TestQueryCounter.__instance is None:
+            TestQueryCounter.__instance = object.__new__(cls)
+            TestQueryCounter.__instance.reset()
+        return TestQueryCounter.__instance
+
+    def new_query(self, sql, params):
+        if "SAVEPOINT" not in sql:
+            self._counter += 1
+            self._queries_stack.append((sql, params))
+
+    def reset(self):
+        self._counter = 0
+        self._queries_stack = []
+
+    def get_counter(self):
+        return self._counter
+
+    def get_queries_stack(self):
+        return self._queries_stack
 
 
 def hacked_execute(self, sql, params=()):
-    global test_query_counter
-    if "SAVEPOINT" not in sql:
-        test_query_counter += 1
+    TestQueryCounter().new_query(sql, params)
     return self.old_execute(sql, params)
 
 
 class query_counter(object):
     def __enter__(self):
         # reset counter
-        global test_query_counter
-        test_query_counter = 0
+        TestQueryCounter().reset()
 
         # patching CursorWrapper
         CursorWrapper.old_execute = CursorWrapper.execute
@@ -35,7 +54,7 @@ class query_counter(object):
         CursorWrapper.execute = CursorWrapper.old_execute
 
         if exc_type is None:
-            global test_query_counter
+            test_query_counter = TestQueryCounter().get_counter()
 
             if test_query_counter > getattr(settings, "TEST_QUERY_NUMBER_RAISE_ERROR", 15):
                 raise TooManySQLQueriesException("Too many queries executed: %d" % test_query_counter)
@@ -65,5 +84,5 @@ class QueryCountingTestCaseMixin(object):
     client_class = QueryCountingAPIClient
 
 
-class QueryCountingApiTestCase(APITestCase, QueryCountingTestCaseMixin):
+class QueryCountingApiTestCase(QueryCountingTestCaseMixin, APITestCase):
     pass