+++ tests/models.py
@@ -34,9 +34,28 @@ class ThirdLevelModelForNestedFilteringTest(models.Model):
 
 class SecondLevelModelForContextPassingTest(models.Model):
     name = models.CharField(max_length=255)
-    third = models.ForeignKey(ThirdLevelModelForNestedFilteringTest, related_name="second", null=True)
+    third = models.ForeignKey(ThirdLevelModelForNestedFilteringTest, related_name="second", null=True,
+                              on_delete=models.CASCADE)
 
 
 class TopLevelModelForContextPassingTest(models.Model):
-    second = models.ForeignKey(SecondLevelModelForContextPassingTest, related_name="top")
+    second = models.ForeignKey(SecondLevelModelForContextPassingTest, related_name="top", on_delete=models.CASCADE)
     name = models.CharField(max_length=255)
+
+
+class AutoOptimization3Model(models.Model):
+    name = models.CharField(max_length=255)
+    sample = models.ForeignKey(SampleModel, on_delete=models.CASCADE)
+
+
+class AutoOptimization2Model(models.Model):
+    name = models.CharField(max_length=255)
+    fk_3_1 = models.ForeignKey(AutoOptimization3Model, related_name="reverse_2_1", on_delete=models.CASCADE)
+    fk_3_2 = models.ForeignKey(AutoOptimization3Model, related_name="reverse_2_2", on_delete=models.CASCADE)
+    sample = models.ForeignKey(SampleModel, on_delete=models.CASCADE)
+
+
+class AutoOptimization1Model(models.Model):
+    name = models.CharField(max_length=255)
+    fk_2 = models.ForeignKey(AutoOptimization2Model, related_name="reverse_1", on_delete=models.CASCADE)
+    sample_m2m = models.ManyToManyField(SampleModel)